Update GitHub CI workflow
This change updates the GitHub Action workflow used for CI, adding
names to the jobs as well as a terminal job that will simplify the
process of adding required status checks to branches.

Having a known terminal job that is consistent across all projects
makes it easy to add a required status check for a branch—one that
will not change as new intermediate jobs are added.

A lint task and the appropriate `yarn` flags have been added as well.