Add 'Contributing' section to README
This section includes the test and release information that was present
already, but adds a "Setup" section to explain how to setup the
development environment. It also explains the `yarn setup` command,
which is a common point of confusion for new contributors.

Linting instructions have been added alongside the test instructions as
well.