Upgrade Jest and use TS for Jest config (#97)
v27 is the latest version of Jest and introduces these major changes:

* `node` is now the default environment instead of `jsdom`. This
  results in faster runtimes.
* The runner has been completely rewritten as `jest-circus` which is
  also the default, replacing `jest-jasmine2`.
  * This should have no impact on us as we don't rely on any
    Jasmine-specific APIs, but is nice to know.
* The "modern" version of Fake Timers (enabled via
  `jest.useFakeTimers("modern")`) is now the default.

More info here: <https://jestjs.io/blog/2021/05/25/jest-27>

Additionally, to ensure that we make use of all of the latest and
greatest, this PR regenerates the Jest config file (as
`jest.config.ts`). This has the added improvement that each option is
now commented so we know exactly what each one does.