Bump prettier-plugin-packagejson to ^2.2.17 (#100)
Although 2.2.17 is a patch bump from 2.2.11,
`prettier-plugin-packagejson` now depends on `sort-package-json` 1.55.0
rather than 1.50.0, which changes how the `scripts` section in
`package.json` is ordered: namely, that it is now alphabetized. This can
be seen in projects that make use of the module template where 2.2.17 of
`prettier-plugin-packagejson `is installed (as it is the latest
version). This commit updates the version to match so that if
`package.json` is updated it is sorted in the same manner.