Extract build-only TypeScript config (#102)
Currently, test files are excluded from TypeScript's purview via the
`exclude` option in `tsconfig.json`. This is helpful when compiling
because it means that test files will not be included in the build
directory. However, for development, this is unnecessary, as we want
TypeScript to "see" and resolve types across all files in the project.

This commit splits off the build-specific settings from `tsconfig.json`
into a `tsconfig.build.json` file, which will exclusively be used in the
`yarn build` script.