Add jest-it-up (#106)
Adds `jest-it-up` to the template. [`jest-it-up`](https://github.com/rbardini/jest-it-up) is a neat package that detects when coverage has increased, and automatically your coverage minimums in the Jest config. The only requirements made of your Jest config are that (1) it is a conventional `jest.config.js` file and (2) its `coverageReporters` includes `json-summary`. Consequently, this PR converts the Jest config file from TypeScript back to plain JavaScript, and adds the required coverage reporter.

I believe that the benefits conferred by `jest-it-up` outweigh those of using TypeScript for the config file, which rarely changes and is never consumed by application code. Meanwhile,  `json-summary` outputs a single JSON file to the `coverage/` directory and does not appear to impact performance.