Update `prettier` to latest version
The package `prettier` and the one Prettier plugin we use have been
updated to their latest versions.

This resolves a CLI warning that used to be emitted upon running the
`lint` command, which stated that the CLI flag
`--no-error-on-unmatched-pattern` was not recognized. This flag wasn't
supported by the older version, but it is supported by this version.