Update action-npm-publish to support Yarn protocols (#144)
The action `MetaMask/action-npm-publish` has been updated to v2, which
uses Yarn for publishing rather than npm. This ensures the publish step
works correctly with Yarn-specific dependency protocols such as
`workspace:`.

The `.yarn-state.yml` file needs to be cached and restored during
publish to ensure that Yarn can run the `prepack` script. The prepack
step is a no-op here because we have `SKIP_PREPACK` set, but there is
no way to tell Yarn to not attempt running it.

Since we're no longer using npm, we don't need to set `ignore-scripts`
in the npm config anymore either.