Reorganize GitHub Action workflows (#149)
* Reorganize GitHub Action workflows

The workflows have been reorganized to use one "main" workflow for any
push to the main branch, with additional jobs being triggered from
there using `workflow_call`. This mirrors the change made recently in
the controllers repository as part of the monorepo branch.

This makes more sense semantically every since we moved the publish
step to be triggered by a push to `main`. Since then, it seend wrong to
call that workflow "build-test" because it also did publishing. Now the
"build-test" workflow is distinct, and it's the "main" workflow that
triggers publishing.

* Remove extra newline

Co-authored-by: legobeat <109787230+legobeat@users.noreply.github.com>

* Fix job name

Co-authored-by: legobeat <109787230+legobeat@users.noreply.github.com>