Improve compatibility of `is-release` job condition (#153)
* Improve compatibility of `is-release` job condition

The `is-release` job condition is now compatible with more release
workflows. Previously it assumed that release commits were made by
GitHub Actions, but that's not true anymore in some of our
repositories.

Instead, we now ensure just that the triggering event was a `push`
event. `push` events are only triggered on this workflow for the main
branch, so that will omit any events from PRs. It also ensures that
the `action-is-release` action is not triggered for `pull_request`
events, which it does not currently support.

`action-is-release` should correctly identify releases, so long as it's
only run on the `main` branch for `push` events in a repository that
only allows `Squash & merge` commits on the main branch. These
requirements have been explained in a comment.

* Fix typo

Co-authored-by: Maarten Zuidhoorn <maarten@zuidhoorn.com>

* Restore github-actions author filter

* improve comment

* Fix workflow name

Co-authored-by: Maarten Zuidhoorn <maarten@zuidhoorn.com>