Fix 'yarn pack'
The `yarn pack` command is run by the `npm-publish` action prior to
publishing the package. Unfortunately, this command does not work as the
`prepack` lifecycle script is overridden to run `yarn build:clean`,
which was removed when this repo was converted to use `tsup` for
building. `tsup` already clears the `dist` directly, so in practice a
`build:clean` script is not usually necessary, but there are occasions
where we want to _really_ be sure that `dist` does not exist, and
publishing is one of them. Hence, this commit adds back `build:clean`,
which should fix `yarn pack` and the publishing workflow.