Fix 'yarn pack' (#238)
The `yarn pack` command is run by the `npm-publish` action prior to
publishing the package. Unfortunately, this command does not work as the
`prepack` lifecycle script is overridden to run `yarn build:clean`,
which was removed when this repo was converted to use `tsup` for
building. `tsup` already has a way of clearing `dist` via a `--clean` flag, so
this commit modifies the `build` script to use it and modifies the prepack
script to run `yarn build` instead of `yarn build:clean`.