Don't cache ESLint results (#295)
The `eslint` executable takes a `--cache` option, which instructs ESLint
to save the results of linting for each file, skipping linting in the
future for files that haven't changed. While this does tend to make
repeated runs faster, in other repos where this technique is also used,
we've noticed that the cache can also give misleading output when
running `eslint` with various other options, making debugging lint
issues frustrating. It is better in the long run to disable caching.